+++ CHANGELOG.md
@@ -1,6 +1,4 @@
-- Fixed issue caused other mods that add new layers to player model to display those incorrectly when River Pikehorn was on player's head
-- Fixed glowing layer not working with shaders on
-- Fixed Swamp Wyvern's easter egg variant "jeb_" not being applicable via nametag
+- Fixed incorrect rotation of River Pikehorn on head
 
 ### Known Issues:
 - Dragon rotations in some scenarios may be very out of sync with the server